+++ sentry-ruby/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Let Transaction constructor take an optional hub argument [#1384](https://github.com/getsentry/sentry-ruby/pull/1384)
 - Introduce LoggingHelper [#1385](https://github.com/getsentry/sentry-ruby/pull/1385)
+- Raise exception if a Transaction is initialized without a hub [#1391](https://github.com/getsentry/sentry-ruby/pull/1391)
 
 ## 4.3.2